centralization
- created 2 buttons on the home page, one that loads ocr and one for chelist
