+++ app/src/main/AndroidManifest.xml
@@ -25,7 +25,6 @@
         <activity android:name=".MainTextRecognizer">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN"/>
-                <category android:name="android.intent.category.LAUNCHER"/>
             </intent-filter>
         </activity>
         <activity
